Move calcDist to distUtils
